Integers fixes (with small modifications), closes #84

+++ src/blackwhitelist.h
@@ -1,2 +1,2 @@
 int blackwhitelist_load_list(const char *filename);
-int blackwhitelist_check_hostname(const char *host_addr, int host_len);
+int blackwhitelist_check_hostname(const char *host_addr, size_t host_len);
